deps/v8/test/inspector/debugger/break-on-exception-expected.txt

Check that inspector correctly change break on exception state.

Running test: noBreakOnExceptionAfterEnabled

Running test: breakOnUncaughtException
paused in throwUncaughtException

Running test: breakOnCaughtException
paused in throwUncaughtException
paused in throwCaughtException

Running test: noBreakInEvaluateInSilentMode
